Require a predis client instead of a client interface.
This was incorrectly typehinting a ClientInterface, but we use methods
that are not available on the interface.  These methods are `get`,
`set`, and `expire`.  Found via scrutinizer:
https://scrutinizer-ci.com/g/dominionenterprises/memoize-php/inspections/4f1bfc08-5879-4f04-9826-cf909d198bed/issues/files/src/Predis.php?status=new&orderField=path&order=asc

We could likely change things to just use the client interface by using
the `createCommand` and `executeCommand`, but for now, I think it's
simpler to just do this and leave that as a future possibility.